+++ CHANGELOG.md
@@ -20,3 +20,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Issue with instance settings allowing setting of memory java params
 
 ### Misc
+- Implement view model for ServersTab [#674]